requirements.txt
torch
torchvision
matplotlib
numpy
Pillow
tqdm
onnx